updated docs link css and component name, moved collection menu to it's own directory
