+++ compositions/CollectionMenu/CollectionMenu.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { Box, Icon, Button, Label, Stack, color } from '@zoralabs/zord'
 import { ModalComposition } from '@modal'
 import { useCollectionsContext } from 'providers/CollectionsProvider'
-import { modalWrapper, collectionTrigger } from '../Header.css'
+import { modalWrapper, collectionTrigger } from './CollectionMenu.css'
 import { noTextWrap } from 'styles/styles.css'
 import { CollectionNavList } from './CollectionNavList'
 import { HorizontalMenu } from 'components'
