+++ compositions/Header/NounsLink.tsx
@@ -1,4 +1,4 @@
-import NextLink from 'next/link'
+import Link from 'next/link'
 import { Box, Flex } from '@zoralabs/zord'
 import { NOUNS_GLASSES } from 'styles/style-constants'
 import { APP_TITLE } from 'utils/env-vars'
@@ -7,10 +7,10 @@ import { clickAnimation } from 'styles/styles.css'
 
 export function NounsLink() {
   return (
-    <NextLink href="/" passHref>
+    <Link href="/" passHref>
       <Flex as="a" className={[clickAnimation, nounsGlassesLink]}>
         <Box as="img" className={[nounsGlasses]} src={NOUNS_GLASSES} alt={APP_TITLE} />
       </Flex>
-    </NextLink>
+    </Link>
   )
 }
